Update 1.4, mostly visual changes.
Fixed some icons spacing problems. As contact-icons, that got it's margin-top removed in the previus version, because the margin was also applying to the icons on footer. Fix solution: adjust the footer's contact icons class to "footer-icons".
Furthermore, I solved the figcaption's problem, which made it look like it wasn't an actual element for the browser. Solution: add a padding-bottom that embraces the figcaption, as if it was an actual part of the figure element (IT IS, BUT FOR SOME REASON THE BROWSER DOESN'T SEE IT AS PART OF IT).
And last (and least, lol), I moved tech icons to a folder called tech-icons, removed items from the font folder that weren't being used.

In short, many changes.